Refactor ArenaWrappedDBIter into separate files
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:
