+++ src.mk
@@ -3,6 +3,7 @@ LIB_SOURCES =                                                   \
   cache/clock_cache.cc                                          \
   cache/lru_cache.cc                                            \
   cache/sharded_cache.cc                                        \
+  db/arena_wrapped_db_iter.cc					\
   db/builder.cc                                                 \
   db/c.cc                                                       \
   db/column_family.cc                                           \
